add return_value and 25line option

+++ ft_printf/libft_utils03.c
@@ -6,7 +6,7 @@
 /*   By: hyudai <hyudai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 11:42:16 by hyudai            #+#    #+#             */
-/*   Updated: 2021/02/06 11:56:09 by hyudai           ###   ########.fr       */
+/*   Updated: 2021/02/06 15:45:32 by hyudai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,11 +25,11 @@ t_poption	printf_struct_reset(t_poption flag)
 
 int		error_handling(t_poption *flag)
 {
-	if (flag->period > 2147483648)
+	if ((long)flag->period > 2147483648)
 		return (-1);
-	if (flag->asterisk > 2147483648)
+	if ((long)flag->asterisk > 2147483648)
 		return (-1);
-	if (flag->number > 2147483648)
+	if ((long)flag->number > 2147483648)
 		return(-1);
 	if (flag->period < 0)
 		flag->period = 0;
